move files with helper functions
