+++ src/components/tabs/get-rating.ts
